test: run zfs test only on freebsd
